feature : '뿌리기' 기능 구현 (뿌린 부분까지)
- 하기 task 에 대한 구현 및 테스트 완료
- SM-API1-S003 : '뿌리기'를 요청한 사용자 자신의 잔액은 '뿌린금액' 만큼 감소되고 '참여대화방'에 뿌리기 메시지가 발송된다.

+++ src/main/java/com/kakaopay/sprinklingmoney/domain/SprinkleMoney.java
@@ -27,7 +27,7 @@ public class SprinkleMoney {
 
     private int sprinkleUserCount;
 
-    private int sprinkleMoneyAmount;
+    private long sprinkleMoneyAmount;
 
     @Enumerated(EnumType.STRING)
     private SprinkleStatus sprinkleStatus;
@@ -46,7 +46,7 @@ public class SprinkleMoney {
     // ===== create method ===== //
     public static SprinkleMoney createSprinkleMoney(UserChatroom userChatroom,
                                                     int sprinkleUserCount,
-                                                    int sprinkleMoneyAmount) {
+                                                    long sprinkleMoneyAmount) {
 
         SprinkleMoney sprinkleMoney = new SprinkleMoney();
